refactor: rename playbook runs page and fix imports
